google-gemini/requirements.txt
langchain
langchain-google-genai
google.cloud
google-cloud-aiplatform
pandas
langchain-community
chromadb
google.generativeai